(GH-22) Move Crash Dump to own module
Previously the Crash Dump module was located in the message_router.  This
commit moves this into it's own module.
